software/embarcado/djangoapp/requirements.txt
Django>=4.0
paho-mqtt==2.1.0
django-environ==0.11.*
requests==2.32.3
djangorestframework==3.15.2
